requirements_cu111.txt
numpy==1.21.0 # OpenPCDet limitations
torch==1.10.1+cu111 --find-links https://download.pytorch.org/whl/cu111/torch_stable.html
torchvision==0.11.2+cu111
spconv-cu111
tensorflow==2.5.0